requirements.txt
openai==0.27.0
tiktoken==0.2.0
python-dotenv
discord.py
duckduckgo_search